requirements.txt
accelerate==1.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
atla==0.6.0
attrs==25.3.0
backoff==2.2.1
blinker==1.9.0
cachetools==5.5.2
cargo==0.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.12
dataclasses-json==0.6.7
datasets==3.5.0
dill==0.3.8
distro==1.9.0
duckduckgo_search==7.5.2
et_xmlfile==2.0.0
eval_type_backport==0.2.2
evaluate==0.4.3
faiss-cpu==1.10.0
filelock==3.17.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.1
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.45
langchain-mistralai==0.2.8
langchain-text-splitters==0.3.6
langfuse==2.60.3
langsmith==0.3.13
lxml==5.3.1
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
mistralai==1.5.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.31.0
networkx==3.4.2
numpy==2.2.3
openpyxl==3.1.5
optimum==1.24.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==11.1.0
primp==0.14.0
propcache==0.3.0
protobuf==5.29.4
psutil==7.0.0
pyarrow==19.0.1
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.9.1
PyMuPDF==1.25.3
pyparsing==3.2.1
pypdf==5.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.23.1
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.39
streamlit==1.43.2
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.6.0
tokenizers==0.21.0
toml==0.10.2
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.48.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
youtube-search==2.1.2
zstandard==0.23.0
